backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.1.4
langchain-community==0.0.16
pytesseract==0.3.10
Pillow==10.2.0
requests==2.31.0
celery==5.3.6
redis==5.0.1
python-socketio==5.11.1
groq 
langchain-groq